chore: refactor validation service
