+++ packages/shared/provider/index.ts
@@ -1,3 +1,2 @@
-export { createProviderService } from './provider-service.js'
-export { createProvider } from './providerFactory.js'
-export { createSignerService } from './signerService.js'
+export * from './provider-service.js'
+export * from './providerFactory.js'
