+++ packages/shared/signer/index.ts
@@ -1,0 +1,1 @@
+export * from './signerService.js'
