--- packages/shared/sim/index.ts
@@ -1,1 +1,0 @@
-export * from './sim.js'
