+++ packages/shared/types/index.ts
@@ -7,4 +7,3 @@ export * from './validation.types.js'
 export * from './bundlerCollectorTracer.types.js'
 export * from './gethTracer.types.js'
 export * from './metrics.types.js'
-export * from './signer.types.js'
