+++ packages/shared/validatation/index.ts
@@ -1,0 +1,1 @@
+export { ValidationService } from './validation-service.js'
