+++ packages/shared/validatation/parseScannerResult.ts
@@ -48,7 +48,7 @@ const abi = Object.values([
  * - last entry is top-level return from "simulateValidation". it as ret and rettype, but no type or address
  * @param tracerResults
  */
-function parseCallStack (tracerResults: BundlerCollectorReturn): CallEntry[] {
+const parseCallStack = (tracerResults: BundlerCollectorReturn): CallEntry[] => {
   const xfaces = new Interface(abi)
 
   function callCatch<T, T1> (x: () => T, def: T1): T | T1 {
@@ -118,10 +118,10 @@ function parseCallStack (tracerResults: BundlerCollectorReturn): CallEntry[] {
  * @param stakeInfoEntities stake info for (factory, account, paymaster). factory and paymaster can be null.
  * @param keccak array of buffers that were given to keccak in the transaction
  */
-function parseEntitySlots (
+const parseEntitySlots = (
   stakeInfoEntities: { [addr: string]: StakeInfo | undefined }, 
   keccak: string[]
-): { [addr: string]: Set<string> } {
+): { [addr: string]: Set<string> } => {
   // for each entity (sender, factory, paymaster), hold the valid slot addresses
   // valid: the slot was generated by keccak(entity || ...)
   const entitySlots: { [addr: string]: Set<string> } = {}
@@ -154,10 +154,10 @@ const callsFromEntryPointMethodSigs: {[key: string]: string} = {
 }
 
 // Helper function to find call info for the entry point
-function findCallInfoEntryPoint(
+const findCallInfoEntryPoint = (
   callStack: CallEntry[],
   entryPointAddress: string
-): CallEntry | undefined {
+): CallEntry | undefined => {
   return callStack.find(
     (call) =>
       call.to === entryPointAddress &&
@@ -168,7 +168,7 @@ function findCallInfoEntryPoint(
 }
 
 // Helper function to find an illegal non-zero value call
-function findIllegalNonZeroValueCall(callStack: CallEntry[], entryPointAddress: string): CallEntry | undefined {
+const findIllegalNonZeroValueCall = (callStack: CallEntry[], entryPointAddress: string): CallEntry | undefined => {
   return callStack.find(
     (call) =>
       call.to !== entryPointAddress &&
@@ -177,12 +177,12 @@ function findIllegalNonZeroValueCall(callStack: CallEntry[], entryPointAddress: 
 }
 
 // Helper function to extract addresses from the call stack
-function extractAddresses(callsFromEntryPoint: TopLevelCallInfo[]): string[] {
+const extractAddresses = (callsFromEntryPoint: TopLevelCallInfo[]): string[] => {
   return callsFromEntryPoint.flatMap((level) => Object.keys(level.contractSize))
 }
 
 // Helper function to extract storage map from the call stack
-function extractStorageMap(callsFromEntryPoint: TopLevelCallInfo[]): StorageMap {
+const extractStorageMap = (callsFromEntryPoint: TopLevelCallInfo[]): StorageMap => {
   const storageMap: StorageMap = {}
 
   callsFromEntryPoint.forEach((level) => {
@@ -202,12 +202,12 @@ function extractStorageMap(callsFromEntryPoint: TopLevelCallInfo[]): StorageMap 
  * @param entryPoint the entryPoint that hosted the "simulatedValidation" traced call.
  * @return list of contract addresses referenced by this UserOp
  */
-export function tracerResultParser(
+export const tracerResultParser = (
   userOp: UserOperation,
   tracerResults: BundlerCollectorReturn,
   validationResult: ValidationResult,
   entryPoint: ethers.Contract
-): [string[], StorageMap] {
+): [string[], StorageMap] => {
   const entryPointAddress = entryPoint.address.toLowerCase()
   
   // banned opcodes from [OP-011]
