--- packages/shared/validation/index.ts
@@ -1,1 +1,0 @@
-export { ValidationService } from './ValidationService.js'
